+++ src/wai/common/meta/exception/__init__.py
@@ -1,0 +1,4 @@
+"""
+Package for working with exceptions.
+"""
+from ._ExceptionChainer import ExceptionChainer